+++ lib/src/controllers/one_notification_controller.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
-import 'package:one_context/src/components/one_notification.dart';
+import 'package:one_context/src/components/one_notifier.dart';
 
 class OneNotificationController {
-  void notify<T>(BuildContext context, [T data]) {
-    OneNotification<T>(data: data)..dispatch(context);
+  void notify<T>(BuildContext context, [NotificationPayload<T> payload]) {
+    OneNotifier.notify<T>(context, payload);
   }
 }